split some slideshow code off into subfolders
