+++ src/Modules/RoomSlideShow/Core/KeyFrame.cs
