+++ src/Modules/RoomSlideShow/Core/Playback.cs
