+++ src/Modules/RoomSlideShow/Core/PlaybackStep.cs
