+++ src/Modules/RoomSlideShow/Core/SlideShowInstant.cs
