+++ src/Modules/RoomSlideShow/SlideShowRectData.cs
@@ -1,0 +1,14 @@
+namespace RegionKit.Modules.Slideshow;
+
+public sealed class SlideShowRectData : ManagedData
+{
+	[StringField("00id", "test", "Id")]
+	public string id = "test";
+	[Vector2Field("01p2", 100f, 100f, Vector2Field.VectorReprType.rect)]
+	public Vector2 p2;
+	public SlideShowRectData(PlacedObject owner) : base(owner, null)
+	{
+
+	}
+
+}
