+++ src/Modules/RoomSlideShow/Steps/EndOfPlayback.cs
