+++ src/Modules/RoomSlideShow/Steps/Frame.cs
