+++ src/Modules/RoomSlideShow/Steps/PlayState.cs
