+++ src/Modules/RoomSlideShow/Steps/SetContainer.cs
