+++ src/Modules/RoomSlideShow/Steps/SetDelay.cs
