+++ src/Modules/RoomSlideShow/Steps/SetInterpolation.cs
