+++ src/Modules/RoomSlideShow/Steps/SetShader.cs
